+++ branches/specialq-nobatch/gnfs/poly/stage1/stage1_core_gpu/stage1_core_sq.h
@@ -43,7 +43,7 @@ typedef struct {
 
 typedef struct {
 	uint32 p[P_SOA_BATCH_SIZE];
-	uint64 lattice_size[P_SOA_BATCH_SIZE];
+	float lattice_size[P_SOA_BATCH_SIZE];
 	uint64 start_root[P_SOA_BATCH_SIZE];
 	uint64 roots[SPECIALQ_BATCH_SIZE][P_SOA_BATCH_SIZE];
 } p_soa_t;
